+++ noodles-fasta/src/fai/record.rs
@@ -11,13 +11,12 @@ const MAX_FIELDS: usize = 5;
 #[derive(Debug, Default, Eq, PartialEq)]
 pub struct Record {
     name: String,
-    len: u64,
+    length: u64,
     offset: u64,
     line_bases: u64,
     line_width: u64,
 }
 
-#[allow(clippy::len_without_is_empty)]
 impl Record {
     /// Creates a FASTA index record.
     ///
@@ -27,13 +26,13 @@ impl Record {
     /// use noodles_fasta::fai;
     /// let record = fai::Record::new("sq0", 8, 4, 80, 81);
     /// ```
-    pub fn new<N>(name: N, len: u64, offset: u64, line_bases: u64, line_width: u64) -> Self
+    pub fn new<N>(name: N, length: u64, offset: u64, line_bases: u64, line_width: u64) -> Self
     where
         N: Into<String>,
     {
         Self {
             name: name.into(),
-            len,
+            length,
             offset,
             line_bases,
             line_width,
@@ -62,8 +61,23 @@ impl Record {
     /// let record = fai::Record::new("sq0", 8, 4, 80, 81);
     /// assert_eq!(record.len(), 8);
     /// ```
+    #[allow(clippy::len_without_is_empty)]
+    #[deprecated(since = "0.23.0", note = "Use `Record::length` instead.")]
     pub fn len(&self) -> u64 {
-        self.len
+        self.length()
+    }
+
+    /// Returns the length of the sequence.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use noodles_fasta::fai;
+    /// let record = fai::Record::new("sq0", 8, 4, 80, 81);
+    /// assert_eq!(record.length(), 8);
+    /// ```
+    pub fn length(&self) -> u64 {
+        self.length
     }
 
     /// Returns the offset from the start.